+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Update checkProof api call
 - Sign envelope as bytes
 - parseProcessResultsDigested -> parseProcessResultsDigestedSingleQuestion, multiQuestion tbd
+- Add pubKeysAreEqual method
 
 ## 0.19.0
 - Update process contract wrappers to smart contracts v2